requirements.txt
pandas
pdfplumber
tabula-py
pyinstaller
-e .